requirements.txt
Django>=2.2.10, < 2.3
psycopg2-binary>=2.8.4, <2.9
djangorestframework>=3.11.0, <3.12

#auth
djoser
djangorestframework-simplejwt>=4.4.0,<4.5

# markdown>=3.1.1,<3.2
# Pygments>=2.5.2,<2.6
# django-filter>=2.2.0,<2.3
# django-guardian>=2.1.0,<2.2
# gunicorn>=20.0.4,<20.1
Faker>=4.0.0,<4.1
django-cors-headers>=3.2.1,<3.3
Pillow>=7.0.0, <7.1 # 图片上传

# coreapi文档
coreapi>=2.3.3, <2.4
pyyaml>=5.3, <5.4

# https from runserver
# django-sslserver